requirements.txt
amqp==5.2.0
appdirs==1.4.3
#apturl==0.5.2
asgiref==3.8.1
attrs==25.3.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==4.2.1
blinker==1.4
#Brlapi==0.7.0
Brotli==1.0.9
CacheControl==0.12.6
celery==5.4.0
certifi==2019.11.28
cffi==1.17.1
channels==4.2.2
chardet==3.0.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
#command-not-found==0.3
configparser==7.1.0
coreapi==2.3.3
coreschema==0.0.4
crayons==0.4.0
cryptography==36.0.2
cssselect==1.1.0
#cupshelpers==1.0
dataclasses-json==0.6.7
dbus-python==1.2.16
#defer==1.0.6
Deprecated==1.2.18
diff-match-patch==20241021
distlib==0.3.8
distro==1.4.0
#distro-info==0.23+ubuntu1.1
Django==4.2.16
django-allauth==65.0.2
django-cors-headers==4.4.0
django-import-export==4.1.1
django-nested-admin==4.1.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-s3direct==2.0.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
dnspython==1.16.0
drf-yasg==1.21.7
#duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.3.0
facebook_page_scraper==5.0.6
fasteners==0.14.1
filelock==3.16.0
Flask==2.0.3
future==0.18.2
#gbinder-python==1.1.2
#gpg==1.13.1
h11==0.13.0
h2==4.1.0
hpack==4.0.0
html5lib==1.0.1
httplib2==0.14.0
hyperframe==6.0.1
idna==2.8
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
kaitaistruct==0.9
keyring==18.0.1
kombu==5.4.2
#language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
ldap3==2.9.1
lockfile==0.12.2
#louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.5
marshmallow==3.22.0
mitmproxy==8.0.0
monotonic==1.5
msgpack==1.0.8
mypy_extensions==1.1.0
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
outcome==1.3.0.post0
packaging==24.1
paramiko==2.6.0
passlib==1.7.4
pexpect==4.6.0
pillow==10.4.0
platformdirs==4.3.2
prompt_toolkit==3.0.48
protobuf==3.19.6
psutil==5.5.1
psycopg2==2.9.9
publicsuffix2==2.20191221
pyasn1==0.6.1
pycairo==1.16.2
pycparser==2.22
pycrypto==2.6.1
pycups==1.9.73
pygltflib==1.16.5
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==2.4.6
pyperclip==1.8.2
pyRFC3339==1.1
PySocks==1.7.1
pytesseract==0.3.13
#python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
#python-debian==0.1.36+ubuntu1.1
python-decouple==3.8
python-monkey-business==1.1.0
pytz==2024.2
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
scour==0.37
SecretStorage==2.3.1
selenium==4.1.0
selenium-wire==5.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==1.9.5
sqlparse==0.5.1
systemd-python==234
tablib==3.5.0
tornado==6.4.2
trimesh==4.7.1
trio==0.27.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
#ubuntu-advantage-tools==8001
#ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.20
urllib3-secure-extra==0.1.0
urwid==2.1.2
usb-creator==0.3.7
vine==5.1.0
virtualenv==20.26.4
wadllib==1.3.3
wcwidth==0.2.13
webdriver-manager==3.2.2
webencodings==0.5.1
Werkzeug==3.0.6
wrapt==1.17.2
wsproto==1.1.0
xkit==0.0.0
zstandard==0.17.0
